Patch 6.1.372 Problem:    With 16 bit ints there are compiler warnings. (Walter Briscoe) Solution:   Change int into long. Files:	    src/structs.h, src/syntax.c

+++ src/structs.h
@@ -592,7 +592,7 @@ struct keyentry
 typedef struct buf_state
 {
     int		    bs_idx;	 /* index of pattern */
-    int		    bs_flags;	 /* flags for pattern */
+    long	    bs_flags;	 /* flags for pattern */
     reg_extmatch_T *bs_extmatch; /* external matches from start pattern */
 } bufstate_T;
